compiler fixes

+++ editor/scimed.cc
@@ -15,6 +15,7 @@
 #include "editor/canvas.h"
 
 #include <iomanip>
+#include <algorithm>
 
 bool
 IsCloseTo(float a, float b, float c = 3)
@@ -173,8 +174,8 @@ void
 DrawLine(const Canvas& canvas, int x, int y, int tx, int ty, ImU32 color)
 {
   ImDrawList* draw_list = ImGui::GetWindowDrawList();
-  const auto  from      = canvas.WorldToScreen(ImVec2{x, y});
-  const auto  to        = canvas.WorldToScreen(ImVec2{tx, ty});
+  const auto  from      = canvas.WorldToScreen(ImVec2{static_cast<float>(x), static_cast<float>(y)});
+  const auto  to        = canvas.WorldToScreen(ImVec2{static_cast<float>(tx), static_cast<float>(ty)});
   draw_list->AddLine(from, to, color);
 }
 
@@ -215,8 +216,8 @@ DrawHorizontalCenteredText(
     int                id)
 {
   const auto size  = ImGui::CalcTextSize(s.c_str());
-  const auto left  = canvas.WorldToScreen(ImVec2{left_p, y_p});
-  const auto right = canvas.WorldToScreen(ImVec2{right_p, y_p});
+  const auto left  = canvas.WorldToScreen(ImVec2{static_cast<float>(left_p), static_cast<float>(y_p)});
+  const auto right = canvas.WorldToScreen(ImVec2{static_cast<float>(right_p), static_cast<float>(y_p)});
   const auto y     = left.y;
   auto       x     = left.x + (right.x - left.x) / 2 - size.x / 2;
   const auto p     = ImVec2{x, y - size.y};
@@ -233,8 +234,8 @@ DrawVerticalCenteredText(
     int                id)
 {
   const auto size   = ImGui::CalcTextSize(s.c_str());
-  const auto top    = canvas.WorldToScreen(ImVec2{x_p, top_p});
-  const auto bottom = canvas.WorldToScreen(ImVec2{x_p, bottom_p});
+  const auto top    = canvas.WorldToScreen(ImVec2{static_cast<float>(x_p), static_cast<float>(top_p)});
+  const auto bottom = canvas.WorldToScreen(ImVec2{static_cast<float>(x_p), static_cast<float>(bottom_p)});
   const auto x      = bottom.x;
   const auto y      = top.y + (bottom.y - top.y) / 2 - size.y / 2;
   const auto p      = ImVec2{x - size.x, y};
@@ -343,7 +344,7 @@ DrawSingleAxisSplits(
   for(auto s : splits)
   {
     line_function(s.position);
-    const auto p = canvas->WorldToScreen(ImVec2{s.position, s.position});
+    const auto p = canvas->WorldToScreen(ImVec2{static_cast<float>(s.position), static_cast<float>(s.position)});
     if(IsCloseTo(coord_function(mouse), coord_function(p)))
     {
       ret = i;
@@ -462,7 +463,7 @@ Scimed::Run(const CanvasConfig& cc, const ScimedConfig& scc)
   auto       tex_id = reinterpret_cast<ImTextureID>(texture->GetId());
   const auto pos    = canvas.WorldToScreen(ImVec2{0, 0});
   const auto size =
-      canvas.WorldToScreen(ImVec2{texture->GetWidth(), texture->GetHeight()});
+      canvas.WorldToScreen(ImVec2{static_cast<float>(texture->GetWidth()), static_cast<float>(texture->GetHeight())});
   draw_list->AddImage(tex_id, pos, size);
 
   const auto current_hover = DrawSplits(scaling.get(), &canvas, scc);
